♻️ refactor: 좋아요 버튼 비즈니스 로직 분리 (#76)
